# [REVIEW] Faster Treelite serialization