--- a/ci/gpu/build.sh
+++ b/ci/gpu/build.sh
@@ -106,7 +106,7 @@ export LD_LIBRARY_PATH=$LD_LIBRARY_PATH_CACHED
 export LD_LIBRARY_PATH_CACHED=""
 
 logger "Install Treelite for GPU testing..."
-python -m pip install -v treelite==0.92rc2 treelite_runtime==0.92rc2
+python -m pip install -v treelite==0.92 treelite_runtime==0.92
 
 cd $WORKSPACE