--- a/cpp/cmake/Dependencies.cmake
+++ b/cpp/cmake/Dependencies.cmake
@@ -126,28 +126,25 @@ set(TREELITE_DIR ${CMAKE_CURRENT_BINARY_DIR}/treelite CACHE STRING
   "Path to treelite install directory")
 ExternalProject_Add(treelite
     GIT_REPOSITORY    https://github.com/dmlc/treelite.git
-    GIT_TAG           6fd01e4f1890950bbcf9b124da24e886751bffe6
+    GIT_TAG           107faa5b63082f5054ae18873a3716e180f64514
     PREFIX            ${TREELITE_DIR}
-    CMAKE_ARGS        -DBUILD_SHARED_LIBS=OFF
-                      -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>
-                      -DENABLE_PROTOBUF=ON
+    CMAKE_ARGS        -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>
                       -DCMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH}
-    BUILD_BYPRODUCTS  ${TREELITE_DIR}/lib/libtreelite.a
-                      ${TREELITE_DIR}/lib/libdmlc.a
-                      ${TREELITE_DIR}/lib/libtreelite_runtime.so
-    UPDATE_COMMAND    ""
-    PATCH_COMMAND     patch -p1 -N < ${CMAKE_CURRENT_SOURCE_DIR}/cmake/treelite_protobuf.patch || true)
+    BUILD_BYPRODUCTS  ${TREELITE_DIR}/src/treelite-build/libtreelite_static.a
+                      ${TREELITE_DIR}/src/treelite-build/_deps/dmlccore-build/libdmlc.a
+                      ${TREELITE_DIR}/src/treelite-build/libtreelite_runtime.so
+    UPDATE_COMMAND    "")
 
 add_library(dmlclib STATIC IMPORTED)
 add_library(treelitelib STATIC IMPORTED)
 add_library(treelite_runtimelib SHARED IMPORTED)
 
 set_property(TARGET dmlclib PROPERTY
-  IMPORTED_LOCATION ${TREELITE_DIR}/lib/libdmlc.a)
+  IMPORTED_LOCATION ${TREELITE_DIR}/src/treelite-build/_deps/dmlccore-build/libdmlc.a)
 set_property(TARGET treelitelib PROPERTY
-  IMPORTED_LOCATION ${TREELITE_DIR}/lib/libtreelite.a)
+  IMPORTED_LOCATION ${TREELITE_DIR}/src/treelite-build/libtreelite_static.a)
 set_property(TARGET treelite_runtimelib PROPERTY
-  IMPORTED_LOCATION ${TREELITE_DIR}/lib/libtreelite_runtime.so)
+  IMPORTED_LOCATION ${TREELITE_DIR}/src/treelite-build/libtreelite_runtime.so)
 
 add_dependencies(dmlclib treelite)
 add_dependencies(treelitelib treelite)