--- a/cpp/include/cuml/ensemble/randomforest.hpp
+++ b/cpp/include/cuml/ensemble/randomforest.hpp
@@ -132,10 +132,7 @@ void print_rf_detailed(const RandomForestMetaData<T, L>* forest);
 template <class T, class L>
 void build_treelite_forest(ModelHandle* model,
                            const RandomForestMetaData<T, L>* forest,
-                           int num_features, int task_category,
-                           std::vector<unsigned char>& data);
-
-std::vector<unsigned char> save_model(ModelHandle model);
+                           int num_features, int task_category);
 
 ModelHandle concatenate_trees(std::vector<ModelHandle> treelite_handles);