## Summary
Speed up serialization of Treelite model objects and reduce overhead in multi-GPU RF prediction.

## Features
* Flat object layout: now Tree is an array of POD objects (Node). As a consequence, all queries for a give node should now be made via methods of Tree class.
* Binary serialization without depending on Protobuf
* Eliminate the use of tempfile.

## Benchmark setup
* Using [this script](https://github.com/hcho3/cuml/tree/rf_benchmark) by @Salonijain27. See the link for instructions.
* AWS EC2 instance g4dn.12xlarge, T4 GPU (16 GB GDDR6) X 4
* Benchmark setting: `n_gpus=2, n_gb=2, n_features=20, depth=25, n_estimators=10`. This leads to a forest consisting of 10 depth-25 trees, and we run through 25 million data rows.

## Benchmark Results

### Aggregate

|Before | After |
|-----|----|
| 269.0 sec | 109.4 sec (**2.46x speedup**) |

### Breakdown by components
Notice that some of the overhead is still yet to be explained. Also note that the actual prediction time is a small portion of the total run time. Due to the nature of distributed algorithm, timing measures are approximate.

**Current cuML**

  | Master | Worker 0 | Worker 1
-- | -- | -- | --
RF->Treelite |   | 32.6 | 32.8
Treelite->Protobuf |   | 10.8 | 11.3
(Copy models workers -> master) |   |   |  
Protobuf->Treelite |   | 26.6 | 23.8
Concatenate Treelite handles | 23.3 |   |  
(Copy model master -> workers) |   |   |  
Protobuf->Treelite |   | 52.2 | 52.3
Treelite->FIL |   | 8.6 | 8.5
FIL predict |   | 2.9 | 1.1

**This PR**

  | Master | Worker 0 | Worker 1
-- | -- | -- | --
RF->Treelite |   | 22.1 | 22.5
Treelite->Bytes |   | < 0.001 | < 0.001
(Copy models workers -> master) |   |   |  
Bytes->Treelite |   | < 0.001 | < 0.001
Concatenate Treelite handles | 3.1 |   |  
(Copy model master -> workers) |   |   |  
Bytes->Treelite |   | < 0.001 | < 0.001
Treelite->FIL |   | 4.8 | 4.7
FIL predict |   | 4.9 | 1.1