@JohnZed 2020-05-14
Very promising! I suspect that the RF -> treelite can be accelerated a lot in a future rev too... I don't know why it should have to be much longer than treelite-> fil in general if we move to convert the representation efficiently.