@hcho3 2020-05-14
> I don't fully get the serialization to frames - looks like it simply passing through the binary format used within tl?

Treelite objects now exposes the Python buffer protocol interface, so that we can transparently convert Treelite objects to memory views with zero overhead. [In `init_from_frames()`](https://github.com/rapidsai/cuml/pull/2263/files#diff-f603c5b28d655eedc8cfc3f7481b1f52R99), we fetch the buffer interface from the Treelite object and cast it into memory view. It is O(1) because it amounts to pointer casting.