@hcho3 2020-05-20
This is quite a strange error:
```


=================================== FAILURES ===================================

_________________________ test_real_algos_runner[FIL] __________________________



algo_name = 'FIL'



    @pytest.mark.parametrize('algo_name', ['UMAP-Supervised',

                                           'DBSCAN',

                                           'LogisticRegression',

                                           'ElasticNet',

                                           'FIL'])

    def test_real_algos_runner(algo_name):

        pair = algorithms.algorithm_by_name(algo_name)

    

        if (algo_name == 'UMAP' and not has_umap()) or \

           (algo_name == 'FIL' and not has_xgboost()):

            pytest.xfail()

    

        runner = AccuracyComparisonRunner(

            [20], [5], dataset_name='classification', test_fraction=0.20

        )

        results = runner.run(pair)[0]

        print(results)

>       assert results["cuml_acc"] is not None

E       KeyError: 'cuml_acc'



cuml/test/test_benchmark.py:190: KeyError
```