@hcho3 2020-05-21
I managed to fix the failing benchmark test. Marking this as ready for review.