@hcho3 2020-05-27
Moving this PR to 0.15 branch. TODO: release latest Treelite to Conda and PyPI.