@JohnZed 2020-06-01
Holding off until conda ready for treelite runtime (shortly)