@hcho3 2020-06-14
List of changes made to Treelite:
* Refactor CMake build; create treelite_runtime Python pkg (dmlc/treelite#167)
* Upgrade to PEP 8 style (dmlc/treelite#175)
* Refactor Python tests with Pytest (dmlc/treelite#176)
* Re-factor scikit-learn converter to use mixins (dmlc/treelite#179)
* Fast binary serialization of Treelite models (dmlc/treelite#178)