@hcho3 2020-06-16
I submitted Treelite 0.92 to conda-forge: https://github.com/conda-forge/staged-recipes/pull/11926. Fingers crossed.