@hcho3 2020-06-16
@JohnZed The Treelite repo contains several round-trip tests for the new serializer: https://github.com/dmlc/treelite/blob/master/tests/cpp/test_serializer.cc