@hcho3 2020-06-16
@jakirkham I addressed all your comments, except the one about casting buffer frames.