@jakirkham 2020-06-16
Thanks for all of the work here @hcho3! 😄 Looks good

Sounds like we are going to handle switching to the `treelite` Conda package in another PR. Is that right?