@hcho3 2020-06-16
> Sounds like we are going to handle switching to the treelite Conda package in another PR. Is that right?

Yes, I’ll work on it after this PR is merged.