@hcho3 2020-06-17
The failing test in the CI should be fixed by #2432